+++ source/Internal/Framework/Module/Facade/ActiveModulesDataProviderBridgeInterface.php
@@ -9,6 +9,7 @@ declare(strict_types=1);
 
 namespace OxidEsales\EshopCommunity\Internal\Framework\Module\Facade;
 
+use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration\Controller;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration\Template;
 
 interface ActiveModulesDataProviderBridgeInterface
@@ -27,4 +28,9 @@ interface ActiveModulesDataProviderBridgeInterface
      * @return Template[][]
      */
     public function getTemplates(): array;
+
+    /**
+     * @return Controller[]
+     */
+    public function getControllers(): array;
 }