+++ tests/Integration/Modules/BaseModuleTestCase.php
@@ -92,10 +92,6 @@ abstract class BaseModuleTestCase extends \OxidEsales\TestingLibrary\UnitTestCas
             );
         }
 
-        if (isset($expectedResult['controllers'])) {
-            $this->assertTrue($validator->checkControllers($expectedResult['controllers']), 'Controllers do not match expectations');
-        }
-
         if (isset($expectedResult['settings_values'])) {
             $this->assertTrue(
                 $validator->checkConfigValues($expectedResult['settings_values']),