+++ tests/Unit/Core/ConfigTest.php
@@ -16,6 +16,8 @@ use OxidEsales\Eshop\Core\Theme;
 use OxidEsales\EshopCommunity\Core\Exception\ExceptionHandler;
 use OxidEsales\EshopCommunity\Core\Registry;
 use OxidEsales\EshopCommunity\Core\ShopIdCalculator;
+use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration\Controller;
+use OxidEsales\EshopCommunity\Internal\Framework\Module\Facade\ActiveModulesDataProviderBridgeInterface;
 use OxidEsales\Facts\Facts;
 use oxRegistry;
 use oxTestModules;
@@ -2380,21 +2382,18 @@ class ConfigTest extends \OxidTestCase
         return $mock;
     }
 
-    /**
-     * Test helper
-     *
-     * @return OxidEsales\EshopCommunity\Core\ModuleControllerProvider mock
-     */
-    private function getModuleControllerMapProviderMock()
+    private function getActiveModulesDataProviderBridgeMock(): ActiveModulesDataProviderBridgeInterface
     {
-        $map = array('cCc' => 'Vendor1\Testmodule\SomeController',
-                     'DDD' => 'Vendor1\OtherTestModule\SomeOtherController',
-                     'eee' => 'Vendor2\OtherTestModule\SomeDifferentController');
+        $map = [
+            new Controller('cCc', 'Vendor1\Testmodule\SomeController'),
+            new Controller('DDD', 'Vendor1\OtherTestModule\SomeOtherController'),
+            new Controller('eee', 'Vendor2\OtherTestModule\SomeDifferentController')
+        ];
 
-        $mock = $this->getMock(\OxidEsales\Eshop\Core\Routing\ModuleControllerMapProvider::class, ['getControllerMap'], [], '', false);
-        $mock->expects($this->any())->method('getControllerMap')->will($this->returnValue($map));
+        $bridge = $this->getMockBuilder(ActiveModulesDataProviderBridgeInterface::class)->getMock();
+        $bridge->method('getControllers')->willReturn($map);
 
-        return $mock;
+        return $bridge;
     }
 
     /**
@@ -2404,7 +2403,7 @@ class ConfigTest extends \OxidTestCase
      */
     private function getControllerClassNameResolverMock()
     {
-        $resolver = oxNew(\OxidEsales\Eshop\Core\Routing\ControllerClassNameResolver::class, $this->getShopControllerMapProviderMock(), $this->getModuleControllerMapProviderMock());
+        $resolver = oxNew(\OxidEsales\Eshop\Core\Routing\ControllerClassNameResolver::class, $this->getShopControllerMapProviderMock(), $this->getActiveModulesDataProviderBridgeMock());
 
         return $resolver;
     }